Running backend/frontend code
Added the client and server folders to the plantparent. The client folder contains the frontend React application aspects and code. The server folder contains the bakcend code that is using an Express.js server entry point, which will be used to serve API requests. Finally there is a rebuild of the json packages with these updates