Calendar functions (#19)
* initialized the calendar

Initialized the calendar functionality to be called within the plant home paging

* Updates to create calendar and display issues

Added a file for the plant profile information holder that will include the calendar function. This file is thus called within the Home page and will display the numerous plants that can added as needed

* Fixed positioning and calendar functions

Added the calendar functions of selecting the day you want to add water or sun and if you want to select all. Also updated the layout to take into account widths and adding/deleting plants

* Updated about page

Updated the contents of the about page with project overview information and sample design