Calendar functions (#41)
* initialized the calendar

Initialized the calendar functionality to be called within the plant home paging

* Updates to create calendar and display issues

Added a file for the plant profile information holder that will include the calendar function. This file is thus called within the Home page and will display the numerous plants that can added as needed

* Fixed positioning and calendar functions

Added the calendar functions of selecting the day you want to add water or sun and if you want to select all. Also updated the layout to take into account widths and adding/deleting plants

* Updated about page

Updated the contents of the about page with project overview information and sample design

* Adding the calendar open function

Adding a function that allows the user to select a day on the calendar and put a time for watering and the amount to water

* Adding chatbot functions

Updated the server to be able to run with chatgpt api to be used within the plant page. The chat will allow the user to get feedback about their plant. Updated styling for the calendar to provide a flip calendar function to allow time delivery of water and amount

* Fixing calendar function overlay

Fixing issues found with the calendar functions overlapping each other and the design being not user friendly visually and functionally

* Updates for plant and chats

Updating the plant layout for each user plant and the functionality. Updated so that the chatbot function is working with the server and can connect and reply to user input

* Fixed home page layout

Added information about how the website works, updated layout designs to fit a more modern look, and added animated, clean text.

* Fixed the about page layout

Added some text updates to the about page as well as updating the function of the links to the different sections of the page

* Updating Calendar Branch

Taking code from the main branch and pushing it to the calendar branch for updating.

* Creation of test login

Creating and adding a login functionality for testing cases. Added authentication class to take user locally store login information and use it for creation of a profile, then the profile will be used for testing of the user functionality